rename plugin name
