--- features/feature-evaluation-plugin/VERSION
@@ -1,1 +1,0 @@
-0.2.2
