--- features/feature-evaluation-plugin/build-docker.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-version=$(<VERSION)
-docker build . -t polusai/feature-evaluation-plugin:${version}
