--- features/feature-evaluation-plugin/src/polus/plugins/features/feature_evaluation/__init__.py
@@ -1,4 +1,0 @@
-"""Feature evaluation plugin."""
-__version__ = "0.2.2"
-from . import features_single
-from . import metrics
