--- features/feature-evaluation-plugin/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Feature Evaluation package."""
