+++ features/feature-segmentation-eval/Dockerfile
@@ -16,5 +16,5 @@ COPY src ${EXEC_DIR}/src
 
 RUN pip3 install ${EXEC_DIR} --no-cache-dir
 
-ENTRYPOINT ["python3", "-m", "polus.plugins.features.feature_evaluation"]
+ENTRYPOINT ["python3", "-m", "polus.plugins.features.feature_segmentation_eval"]
 CMD ["--help"]
