+++ features/feature-segmentation-eval/README.md
@@ -1,4 +1,4 @@
-# Feature Evaluation Plugin (v0.2.2)
+# Feature segmentation eval (v0.2.3)
 
 Plugin to generate evaluation metrics for feature comparison of ground truth and predicted images. Contact [Vishakha Goyal](mailto:vishakha.goyal@nih.gov), [Hamdah Shafqat Abbasi](mailto:hamdahshafqat.abbasi@nih.gov) for more information.
 
@@ -15,7 +15,7 @@ If WIPP is running, navigate to the plugins page and add a new plugin. Paste the
 
 ## Options
 
-This plugin takes four input arguments and one output argument:
+This plugin takes six input arguments and one output argument:
 
 | Name          | Description             | I/O    | Type   |
 |---------------|-------------------------|--------|--------|
