+++ features/feature-segmentation-eval/build-docker.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+version=$(<VERSION)
+docker build . -t polusai/feature-segmentation-eval:${version}
