+++ features/feature-segmentation-eval/pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
-name = "polus-plugins-features-feature-evaluation"
-version = "0.2.2"
-description = "Feature Evaluation Plugin"
+name = "polus-plugins-features-feature-segmentation-eval"
+version = "0.2.3"
+description = "Feature segmentation eval"
 authors = [
            "Vishakha Goyal <vishakha.goyal@nih.gov>",
            "Hamdah Shafqat Abbasi <hamdahshafqat.abbasi@nih.gov>"
@@ -21,6 +21,14 @@ blake3 = "^0.3.3"
 llvmlite = "^0.39.1"
 fastapi = "^0.92.0"
 vaex = "^4.7.0"
+
+[tool.poetry.group.dev.dependencies]
+bump2version = "^1.0.1"
+pre-commit = "^3.1.0"
+black = "^23.1.0"
+flake8 = "^6.0.0"
+mypy = "^1.0.1"
+pytest = "^7.2.1"
 
 [build-system]
 requires = ["poetry-core"]
