+++ features/feature-segmentation-eval/run-plugin.sh
@@ -10,10 +10,10 @@ file_pattern=".*.csv"
 out_dir=${datapath}/output
 
 # #Show the help options
-# #docker run polusai/feature-evaluation-plugin:${version}
+# #docker run polusai/feature-segmentation-eval:${version}
 
 docker run -v ${datapath}:${datapath} \
-            polusai/feature-evaluation-plugin:${version} \
+            polusai/feature-segmentation-eval:${version} \
             --GTDir ${gt_dir} \
             --PredDir ${pred_dir} \
             --filePattern ${file_pattern} \
