+++ features/feature-segmentation-eval/src/polus/plugins/features/feature_segmentation_eval/__init__.py
@@ -1,0 +1,4 @@
+"""Feature segmentation evaluation package."""
+__version__ = "0.2.2"
+from . import feature_evaluation
+from . import metrics
