+++ features/feature-segmentation-eval/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Feature segmentation evaluation package."""
